requirements.txt
anyio==4.8.0
Brotli==1.1.0
certifi==2025.1.31
click==8.1.8
duckduckgo_search==7.4.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
lxml==5.3.1
pillow==11.1.0
sniffio==1.3.1
socksio==1.0.0
